backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
pydantic==2.11.4
starlette==1.0.0
anyio==4.9.0
